pyproject.toml
[project]
name = "realestate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.5",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "databases>=0.9.0",
    "dependency-injector>=4.46.0",
    "fastapi-cache2>=0.2.2",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "langchain>=0.3.23",
    "langchain-mcp-adapters>=0.0.7",
    "langchain-openai>=0.3.8",
    "langgraph>=0.3.11",
    "langgraph-checkpoint-mongodb>=0.1.3",
    "langgraph-checkpoint-redis>=0.0.8",
    "langmem>=0.0.28",
    "logger>=1.4",
    "mcp[cli]>=1.5.0",
    "motor>=3.7.0",
    "nest-asyncio>=1.6.0",
    "openai-agents>=0.0.6",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "pydantic-ai>=0.0.43",
    "pydantic-ai-slim[openai]>=0.0.43",
    "pymongo>=4.11.3",
    "pymongoarrow>=1.7.1",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]
